requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.6.0
